+++ src/chains/RobinHood/Common.sol
@@ -10,6 +10,7 @@ import {IPoolManager} from "../../core/UniswapV4Types.sol";
 import {EkuboV3} from "../../core/EkuboV3.sol";
 import {Hanji} from "../../core/Hanji.sol";
 import {Bebop} from "../../core/Bebop.sol";
+import {FluxPool} from "../../core/FluxPool.sol";
 
 import {ISettlerActions} from "../../ISettlerActions.sol";
 import {ISignatureTransfer} from "@permit2/interfaces/ISignatureTransfer.sol";
@@ -46,7 +47,16 @@ import {FastLogic} from "../../utils/FastLogic.sol";
 import {SettlerSwapAbstract} from "../../SettlerAbstract.sol";
 import {Permit2PaymentAbstract} from "../../core/Permit2PaymentAbstract.sol";
 
-abstract contract RobinHoodMixin is FreeMemory, SettlerBase, UniswapV4, EkuboV3, Hanji, PancakeInfinity, Bebop {
+abstract contract RobinHoodMixin is
+    FreeMemory,
+    SettlerBase,
+    UniswapV4,
+    EkuboV3,
+    Hanji,
+    PancakeInfinity,
+    Bebop,
+    FluxPool
+{
     using FastLogic for bool;
 
     constructor() {
@@ -96,6 +106,11 @@ abstract contract RobinHoodMixin is FreeMemory, SettlerBase, UniswapV4, EkuboV3,
             ) = abi.decode(data, (IERC20, uint256, address, uint256, uint256, bool, uint256, uint256));
 
             sellToHanji(sellToken, ppm, pool, sellScalingFactor, buyScalingFactor, isAsk, priceLimit, minBuyAmount);
+        } else if (action == uint32(ISettlerActions.FLUXPOOL.selector)) {
+            (IERC20 sellToken, uint256 ppm, bytes32 poolId, bool zeroForOne, uint256 minBuyAmount) =
+                abi.decode(data, (IERC20, uint256, bytes32, bool, uint256));
+
+            sellToFluxPool(sellToken, ppm, poolId, zeroForOne, minBuyAmount);
         } else if (action == uint32(ISettlerActions.BEBOP.selector)) {
             (
                 address recipient,
@@ -159,6 +174,10 @@ abstract contract RobinHoodMixin is FreeMemory, SettlerBase, UniswapV4, EkuboV3,
         } else {
             revertUnknownForkId(forkId);
         }
+    }
+
+    function _fluxSwap() internal pure override returns (address) {
+        return 0x388e0D8f610a80C75e8a049C39DDc5816f56c012;
     }
 
     function _POOL_MANAGER() internal pure override returns (IPoolManager) {